:truck: [#4246] Move schema assert to shared utility
So that this can be re-used in the digid_eherkenning_oidc package
